Added match deletion warning.

+++ ui/src/components/matches/match_list.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from "react";
 import Container from "react-bootstrap/Container";
-import {Button, Table} from "react-bootstrap";
+import {Button, Modal, Table} from "react-bootstrap";
 import {deleteMatch, fetchMatches, fetchMatchOutcomes, getMatchCompletion} from "./match_service";
 import {FaPlay, FaTrash} from "react-icons/fa";
 import {Link} from "react-router-dom";
@@ -30,11 +30,33 @@ export default function MatchList() {
         })()
     }, [matches]);
 
-    const buildMatchDeleteClickHandler = (matchId: string) => {
-        return async (e: any) => {
+    const defaultDeleteMatchModalProps: DeleteMatchModalProps = {
+        show: false,
+        onClose: null,
+        match: null,
+    }
+    const [deleteMatchModalProps, setDeleteMatchModalprops] = useState(defaultDeleteMatchModalProps)
+
+    const handleMatchDeleteClose = async (matchId: string, deleteConfirmed: boolean) => {
+        if (deleteConfirmed) {
             await deleteMatch(matchId)
             const matches = await fetchMatches()
             setMatches(matches)
+        }
+
+        setDeleteMatchModalprops(
+            {...deleteMatchModalProps, show: false}
+        )
+    }
+
+    const buildMatchDeleteClickHandler = (matchId: string) => {
+        return (e: any) => {
+            const match = matches.filter(
+                m => m.id == matchId
+            )[0]
+            setDeleteMatchModalprops(
+                {...deleteMatchModalProps, match: match, show: true}
+            )
         }
     }
 
@@ -43,7 +65,6 @@ export default function MatchList() {
         const createdAt = match.created_at != null ? match.created_at.toLocaleString() : 'N/A'
         rows.push(
             <tr key={match.id}>
-                <td>{match.id}</td>
                 <td>{match.name}</td>
                 <td>{createdAt}</td>
                 <td><ProgressCounter status={statuses[match.id]}/></td>
@@ -51,7 +72,7 @@ export default function MatchList() {
                     <Link className="m-1" to={`/matches/${match.id}`}>
                         <Button><FaPlay/></Button>
                     </Link>
-                    <Button  className="m-1" variant="danger" onClick={buildMatchDeleteClickHandler(match.id)}><FaTrash/></Button>
+                    <Button className="m-1" variant="danger" onClick={buildMatchDeleteClickHandler(match.id)}><FaTrash/></Button>
                 </td>
             </tr>
         )
@@ -64,9 +85,8 @@ export default function MatchList() {
                 <Table>
                     <thead>
                     <tr>
-                        <th>#</th>
                         <th>Назва</th>
-                        <th>Дата</th>
+                        <th>Створено</th>
                         <th>Статус</th>
                         <th>Дії</th>
                     </tr>
@@ -76,6 +96,46 @@ export default function MatchList() {
                     </tbody>
                 </Table>
             </Container>
+            <DeleteMatchModal
+                match={deleteMatchModalProps.match}
+                show={deleteMatchModalProps.show}
+                onClose={handleMatchDeleteClose}
+            />
         </>
+    )
+}
+
+interface DeleteMatchModalProps {
+    match: MatchInProgress
+    show: boolean
+    onClose: (matchId: string, deleteConfirmed: boolean) => void
+}
+
+function DeleteMatchModal(props: DeleteMatchModalProps) {
+    if (props.match == null) {
+        return <></>
+    }
+    return (
+        <Modal show={props.show} onHide={() => props.onClose(props.match.id, false)}>
+            <Modal.Header closeButton>
+                <Modal.Title>Архівувати матч {props.match.name}?</Modal.Title>
+            </Modal.Header>
+            <Modal.Footer>
+                <Container className={"d-flex justify-content-between"}>
+                    <Button
+                        onClick={event => props.onClose(props.match.id, false)}
+                        variant={"outline-dark"}>
+                        <FaTrash/> Закрити
+                    </Button>
+
+                    <Button
+                        onClick={event => props.onClose(props.match.id, true)}
+                        variant={"danger"}>
+                        <FaTrash/> Архівувати
+                    </Button>
+
+                </Container>
+            </Modal.Footer>
+        </Modal>
     )
 }
